Rename api types file
