--- src/client/api/package.json
@@ -1,26 +1,0 @@
-{
-    "name": "@vscode/pythonExtensionApi",
-    "description": "VSCode Python extension's public API",
-    "version": "1.0.0",
-    "publisher": "ms-python",
-    "author": {
-      "name": "Microsoft Corporation"
-    },
-    "types": "./index.d.ts",
-    "license": "MIT",
-    "homepage": "https://github.com/Microsoft/vscode-python",
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/Microsoft/vscode-python"
-    },
-    "bugs": {
-        "url": "https://github.com/Microsoft/vscode-python/issues"
-    },
-    "dependencies": {
-        "@types/vscode": "^1.78.0"
-    },
-    "devDependencies": {
-        "@types/node": "^16.11.7",
-        "typescript": "^4.7.2"
-    }
-}
